+++ packages/common/src/projects/defaults.ts
@@ -1,4 +1,4 @@
-import { IHubProject } from "../core";
+import { IHubProject, PROJECT_STATUSES } from "../core";
 import { IModel } from "../types";
 
 export const HUB_PROJECT_ITEM_TYPE = "Hub Project";
@@ -10,7 +10,7 @@ export const DEFAULT_PROJECT: Partial<IHubProject> = {
   name: "No title provided",
   tags: [],
   typeKeywords: ["Hub Project"],
-  status: "inactive",
+  status: PROJECT_STATUSES.notStarted,
   catalog: { schemaVersion: 0 },
   permissions: [],
   schemaVersion: 1,
@@ -23,7 +23,7 @@ export const DEFAULT_PROJECT_MODEL: IModel = {
   item: {
     type: HUB_PROJECT_ITEM_TYPE,
     title: "No Title Provided",
-    description: "No Description Provided",
+    description: "",
     snippet: "",
     tags: [],
     typeKeywords: ["Hub Project"],
@@ -35,7 +35,7 @@ export const DEFAULT_PROJECT_MODEL: IModel = {
   data: {
     display: "about",
     timeline: {},
-    status: "inactive",
+    status: PROJECT_STATUSES.notStarted,
     contacts: [],
   },
 } as unknown as IModel;