+++ packages/common/src/search/types/types.ts
@@ -1,3 +1,5 @@
+import { EntityType, IHubCatalog } from "./IHubCatalog";
+
 /**
  * Sort Option to be displayed in a UI
  */
@@ -81,4 +83,51 @@ export interface IApiDefinition {
   url: string;
   // We can add types as we add support for more
   type: "arcgis" | "arcgis-hub";
+}
+/**
+ * Base options when checking catalog containment
+ */
+export interface IContainsOptions {
+  /**
+   * Entity type of the identifier
+   * Specifing this will allow us to skip a lookup
+   */
+  entityType?: EntityType;
+}
+
+/**
+ * Basic information about a catalog
+ */
+export interface ICatalogInfo extends Partial<IDeepCatalogInfo> {}
+
+/**
+ * Cacheable information about a catalog
+ */
+export interface IDeepCatalogInfo {
+  // id of the entity with the catalog we are checking
+  id: string;
+  // Entity type that the catalog belongs to. This enables us to
+  // fetch the catalog.
+  entityType: EntityType;
+  // optional, but if passed it reduces the xhrs to fetch the catalogs
+  catalog?: IHubCatalog;
+}
+
+/**
+ * When checking containment, we want to be able to cache the response
+ * so we return enough information to do that
+ */
+export interface IContainsResponse {
+  // identifier being checked
+  identifier: string;
+  // is the entity actually contained in the catalog?
+  isContained: boolean;
+  /**
+   * Cacheable information about the catalogs
+   */
+  catalogInfo?: Record<string, ICatalogInfo>;
+  /**
+   * How long did it take to check containment?
+   */
+  duration?: number;
 }