+++ packages/common/src/utils/index.ts
@@ -1,3 +1,4 @@
+export * from "./asyncForEach";
 export * from "./batch";
 export * from "./ensure-unique-string";
 export * from "./create-operation-pipeline";