+++ packages/common/test/items/create-item-from-file.test.ts
@@ -22,6 +22,10 @@ describe("createItemFromFile", () => {
         text: "This is a test",
         async: false,
         access: "private",
+        extent: [
+          [1, 2],
+          [3, 4],
+        ],
         file: new Blob(["foo"], { type: "csv" }),
       } as IItemAdd;
       // spies
@@ -47,6 +51,57 @@ describe("createItemFromFile", () => {
       expect(addItemPartSpy).toHaveBeenCalledTimes(3);
       expect(createItemSpy).toHaveBeenCalledTimes(1);
       expect(commitItemUploadSpy).toHaveBeenCalledTimes(1);
+      expect(commitItemUploadSpy.calls.argsFor(0)[0].item.extent).toEqual(
+        "1, 2, 3, 4"
+      );
+      expect(_prepareUploadRequestsSpy).toHaveBeenCalledTimes(1);
+      expect(cancelItemSpy).not.toHaveBeenCalled();
+    });
+    it("Properly creates item w/ string extent", async () => {
+      // request options
+      const ro = {
+        authentication: {
+          portal: "http://some-org.mapsqaext.arcgis.com",
+        },
+      } as IUserRequestOptions;
+      // fake item
+      const item = {
+        title: "Test.csv",
+        type: "csv",
+        owner: "test",
+        dataUrl: "https://test.com",
+        text: "This is a test",
+        async: false,
+        access: "private",
+        extent: "1, 2, 3, 4",
+        file: new Blob(["foo"], { type: "csv" }),
+      } as unknown as IItemAdd;
+      // spies
+      const createItemSpy = spyOn(portal, "createItem").and.returnValue(
+        Promise.resolve({ id: "123abc", success: true, folder: "test" })
+      );
+      const cancelItemSpy = spyOn(portal, "cancelItemUpload").and.returnValue(
+        Promise.resolve({ success: true, id: "123abc" })
+      );
+      const addItemPartSpy = spyOn(portal, "addItemPart").and.callFake(() =>
+        Promise.resolve({ success: true })
+      );
+      const commitItemUploadSpy = spyOn(
+        portal,
+        "commitItemUpload"
+      ).and.returnValue(Promise.resolve({ success: true, id: "123abc" }));
+      const _prepareUploadRequestsSpy = spyOn(
+        _prepareUploadRequestsModule,
+        "_prepareUploadRequests"
+      ).and.returnValue([{}, {}, {}]);
+      const result = await createItemFromFile(item, ro);
+      expect(result).toEqual({ id: "123abc", success: true, folder: "test" });
+      expect(addItemPartSpy).toHaveBeenCalledTimes(3);
+      expect(createItemSpy).toHaveBeenCalledTimes(1);
+      expect(commitItemUploadSpy).toHaveBeenCalledTimes(1);
+      expect(commitItemUploadSpy.calls.argsFor(0)[0].item.extent).toEqual(
+        "1, 2, 3, 4"
+      );
       expect(_prepareUploadRequestsSpy).toHaveBeenCalledTimes(1);
       expect(cancelItemSpy).not.toHaveBeenCalled();
     });