+++ packages/initiatives/CHANGELOG.md
@@ -1,3 +1,68 @@
+## @esri/hub-initiatives [11.16.1](https://github.com/Esri/hub.js/compare/@esri/hub-initiatives@11.16.0...@esri/hub-initiatives@11.16.1) (2022-09-29)
+
+
+
+
+
+### Dependencies
+
+* **@esri/hub-common:** upgraded to 11.16.1
+
+# @esri/hub-initiatives [11.16.0](https://github.com/Esri/hub.js/compare/@esri/hub-initiatives@11.15.0...@esri/hub-initiatives@11.16.0) (2022-09-29)
+
+
+
+
+
+### Dependencies
+
+* **@esri/hub-common:** upgraded to 11.16.0
+
+# @esri/hub-initiatives [11.15.0](https://github.com/Esri/hub.js/compare/@esri/hub-initiatives@11.14.2...@esri/hub-initiatives@11.15.0) (2022-09-29)
+
+
+
+
+
+### Dependencies
+
+* **@esri/hub-common:** upgraded to 11.15.0
+
+## @esri/hub-initiatives [11.14.2](https://github.com/Esri/hub.js/compare/@esri/hub-initiatives@11.14.1...@esri/hub-initiatives@11.14.2) (2022-09-28)
+
+
+
+
+
+### Dependencies
+
+* **@esri/hub-common:** upgraded to 11.14.2
+
+## @esri/hub-initiatives [11.14.1](https://github.com/Esri/hub.js/compare/@esri/hub-initiatives@11.14.0...@esri/hub-initiatives@11.14.1) (2022-09-27)
+
+
+### Bug Fixes
+
+* bump arcgis-rest-portal to pull in request fix for extents ([#895](https://github.com/Esri/hub.js/issues/895)) ([3494865](https://github.com/Esri/hub.js/commit/349486574fb15e47f025c3573578a3ec2719005c))
+
+
+
+
+
+### Dependencies
+
+* **@esri/hub-common:** upgraded to 11.14.1
+
+# @esri/hub-initiatives [11.14.0](https://github.com/Esri/hub.js/compare/@esri/hub-initiatives@11.13.0...@esri/hub-initiatives@11.14.0) (2022-09-26)
+
+
+
+
+
+### Dependencies
+
+* **@esri/hub-common:** upgraded to 11.14.0
+
 # @esri/hub-initiatives [11.13.0](https://github.com/Esri/hub.js/compare/@esri/hub-initiatives@11.12.0...@esri/hub-initiatives@11.13.0) (2022-09-26)